+++ android/src/main/kotlin/com/example/dgis_flutter/CustomLocationManager.kt
@@ -56,6 +56,7 @@ class CustomLocationManager(private val applicationContext: Context): LocationSo
         }
         val request = LocationRequest.create().apply {
             priority = newPriority
+            interval = 100L
         }
         val callback = (object: LocationCallback() {
             override fun onLocationResult(result: LocationResult?) {
@@ -72,6 +73,7 @@ class CustomLocationManager(private val applicationContext: Context): LocationSo
         }
         client.requestLocationUpdates(request, callback, Looper.getMainLooper())
         listener.onAvailabilityChanged(true)
+
     }
 
     private fun checkPermission(): Boolean = ActivityCompat.checkSelfPermission(